+++ CHANGES.md
@@ -37,6 +37,8 @@
 - `plot_feat` used to coerce NaNs and Infs via `np.nan_to_num` prior to plotting, potentially impacting distributions, plotting scales, moments, etc. Fixed so that nan and inf values are removed rather than coerced.
 - Fixed early-stopping statement in `fold_train_ensemble` to state the number as "sub-epochs" (previously said "epochs")
 - Fixed error in patience when using cyclical LR callbacks, now specify the number of cycles to go without improvement. Previously had to specify 1+number.
+- Unnecessary warning `df2foldfile` when no strat-key is passed.
+- Saved matrices in `fold2foldfile` are now in float32.
 
 ## Changes